Separated Cloud Function implementation from 'index.html'; uid added to logging if provided
